requirements.txt
streamlit>=1.24.0
python-dotenv>=0.21.0
crewai==0.28.6
langchain>=0.1.10,<0.2.0
langchain-community>=0.0.9
langchain-openai>=0.0.5
yfinance>=0.2.18
pandas>=1.3.5,<2.0.0
numpy>=1.22.0
matplotlib>=3.5.1
plotly>=5.10.0
requests>=2.28.0
# DeepSeek相关依赖
# 更新于 2025年 5月 9日 星期五 23时53分27秒 CST
chromadb==0.3.26
# 更新于 2025年 5月10日 星期六 00时17分08秒 CST - 添加兼容性 chromadb 版本
pysqlite3-binary>=0.5.0
crewai-tools>=0.1.6
# 最后更新于 2024年5月10日
